requirements.txt
opencv-contrib-python
numpy
Flask
RPi.GPIO
Pillow